Adiciona paginação e ordenação aos controladores de ClientesBanco, Transacoes e Usuarios

+++ src/main/java/com/amandaramos/service/impl/UsuariosServiceImp.java
@@ -3,10 +3,15 @@ package com.amandaramos.service.impl;
 import com.amandaramos.dto.UsuariosDTO;
 import com.amandaramos.entity.Usuarios;
 import com.amandaramos.repository.UsuariosRepository;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -27,13 +32,14 @@ public class UsuariosServiceImp implements UsuarioServiceInterface {
         return convertToDTO(savedUsuario);
     }
 
-    public List<UsuariosDTO> buscarTodosUsuarios() {
-        List<Usuarios> usuarios = usuariosRepository.findAll();
-        return usuarios.stream().map(this::convertToDTO).collect(Collectors.toList());
+    public Page<UsuariosDTO> buscarTodosUsuariosPaginado(Pageable pageable) {
+        Page<Usuarios> usuariosPage = usuariosRepository.findAll(pageable);
+        return usuariosPage.map(this::convertToDTO);
     }
 
     public UsuariosDTO atualizarUsuario(Long id, UsuariosDTO usuariosDTOAtualizado) {
-        Usuarios usuarios = usuariosRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Usuario não encontrado com ID: " + id));
+        Usuarios usuarios = usuariosRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("Usuário não encontrado com ID: " + id));
         // Atualize os campos do usuário com os dados do DTO atualizado
         usuarios.setUsername(usuariosDTOAtualizado.getUsername());
         usuarios.setSenha(usuariosDTOAtualizado.getSenha());
@@ -42,14 +48,17 @@ public class UsuariosServiceImp implements UsuarioServiceInterface {
     }
 
     public Optional<UsuariosDTO> buscarUsuarioPorId(Long id) {
-        Optional<Usuarios> usuarioOptional = usuariosRepository.findAllById(id);
+        Optional<Usuarios> usuarioOptional = usuariosRepository.findById(id);
         return usuarioOptional.map(this::convertToDTO);
     }
 
     public boolean deletarUsuarioPorId(Long id) {
-
-        usuariosRepository.deleteById(id);
-        return false;
+        if (usuariosRepository.existsById(id)) {
+            usuariosRepository.deleteById(id);
+            return true; // Retorna true se o usuário foi deletado com sucesso
+        } else {
+            return false; // Retorna false se o usuário não foi encontrado
+        }
     }
 
     private UsuariosDTO convertToDTO(Usuarios usuario) {
